+++ apps/simulator/src/components/blocks/CaseStudy/CaseStudyQuestions/DND/DNDBlock.tsx
@@ -1,0 +1,20 @@
+import React from 'react'
+import { DND } from '@/components/blocks/CaseStudy/CaseStudyQuestions/DND/DND';
+import { DNDValidationAtom } from '@/core/schema/useAtomic';
+import { DNDValidationType } from '@/core/schema/dnd/validation';
+import { SsrData } from '@/core/types/ssrData'
+import { useAtom } from 'jotai';
+
+export const DNDBlock: React.FC<SsrData> = ({ questionaire }) => {
+    const [dndAtom, setDndAtom] = useAtom(DNDValidationAtom);
+
+    async function handleSubmit(value: DNDValidationType) {
+        console.log("submit", value);
+    }
+
+    return <DND
+        handleSubmit={handleSubmit}
+        dndAtom={dndAtom}
+        questionaire={questionaire} />
+
+}
