+++ apps/simulator/src/components/blocks/CaseStudy/CaseStudyQuestions/index.ts
@@ -5,8 +5,6 @@ export * from './DDTable/DDTQuestionaireBlock'
 export * from './MRSN/MRSNBlock'
 export * from './MCQCSQuestionnaire'
 export * from './MCQCSQuestionnaire'
-export * from './DNDQuestionaire'
+export * from './DND/DNDBlock'
 export * from './MCQTable/MCQAnswerGroupTable'
 export * from './DNDBowtieQuestionaire'
-
-
